add threeDimensionPolygon

+++ src/main/java/jp/abyss/spigot/plugin/customparticle/core/DrawCircleAble.java
@@ -5,7 +5,7 @@ import org.bukkit.Particle;
 
 public interface DrawCircleAble extends DrawThreeDimensionCircleAble {
 
-    void drawCircle(Particle particle, Location center,double radius,double quantity);
+    void drawCircle(Particle particle, Location center,double radius,int quantity);
 
     void drawCircle(Particle particle,Location center,double radius);
