+++ src/main/java/jp/abyss/spigot/plugin/customparticle/core/DrawThreeDimensionCircleAble.java
@@ -4,5 +4,5 @@ import org.bukkit.Location;
 import org.bukkit.Particle;
 
 public interface DrawThreeDimensionCircleAble extends DrawPointAble{
-    void drawThreeDimensionCircle(Particle particle, Location location,double radius,double point);
+    void drawThreeDimensionCircle(Particle particle, Location location,double radius,int point);
 }
